When using Syskit the task sometimes sends the inital command to the servo  (start hook) but the command is not sent to the driver.
If the initial command is not sent the task won't do much and won't prompt
 any error.

I moved the sent of the initial command to the handle sweep method so that
this command is only sent when for sure the connections have been
established.

As the task works on scripts (AFAIK) I would rather not push this changes
to master. Maybe there is a nicer way to check that the dynamixel
controller task will receive the command before sending it.